test(engine-server): reorganize e2e tests
