+++ packages/engine-server/tests/e2e/content-invalid-schema.test.ts
@@ -1,0 +1,31 @@
+import { assert, test } from 'vitest'
+import { createSchema, SchemaDefinition as def } from '@contember/schema-definition'
+import { createTester, gql } from '../src/tester'
+
+
+namespace TagModel {
+	export class Tag {
+		label = def.stringColumn()
+	}
+}
+
+test('Content API: invalid schema error', async () => {
+	const tester = await createTester(createSchema(TagModel))
+
+	await tester(
+		gql`
+			mutation {
+				createFoo(data: { label: "graphql" }) {
+					ok
+				}
+			}
+		`,
+	)
+		.expect(400)
+		.expect(response => {
+			assert.deepStrictEqual(
+				response.body.errors[0].message,
+				'Cannot query field "createFoo" on type "Mutation". Did you mean "createTag"?',
+			)
+		})
+})
