+++ packages/engine-server/tests/e2e/homepage.test.ts
@@ -1,0 +1,10 @@
+import { test } from 'vitest'
+import supertest from 'supertest'
+import { apiUrl } from '../src/tester'
+
+test('show homepage', async () => {
+	await supertest(apiUrl)
+		.get('/')
+		.expect(200)
+		.expect('App is running')
+})
