+++ packages/engine-server/tests/e2e/x-contember-ref.test.ts
@@ -1,0 +1,71 @@
+import { assert, test } from 'vitest'
+import { createTester, gql } from '../src/tester'
+import { createSchema, SchemaDefinition as def } from '@contember/schema-definition'
+
+namespace TagModel {
+	export class Tag {
+		label = def.stringColumn()
+	}
+}
+
+
+test('Content API: X-Contember-Ref header', async () => {
+
+	const tester = await createTester(createSchema(TagModel))
+
+	await tester(
+		gql`
+			mutation {
+				createTag(data: { label: "typescript" }) {
+					ok
+				}
+			}
+		`,
+	).expect(200)
+
+	const response = await tester(
+		gql`
+			query {
+				listTag(filter: { label: { eq: "typescript" } }) {
+					label
+				}
+			}
+		`,
+	)
+		.set('X-Contember-Ref', 'None')
+		.expect(response => {
+			assert.deepStrictEqual(response.body.data, {
+				listTag: [
+					{
+						label: 'typescript',
+					},
+				],
+			})
+		})
+		.expect(200)
+	const eventKey = response.get('X-Contember-Ref')
+	await tester(
+		gql`
+			query {
+				listTag {
+					label
+				}
+			}
+		`,
+	)
+		.set('X-Contember-Ref', eventKey)
+		.expect(304)
+
+	// ignored for mutation
+	await tester(
+		gql`
+			mutation {
+				createTag(data: { label: "typescript" }) {
+					ok
+				}
+			}
+		`,
+	)
+		.set('X-Contember-Ref', eventKey)
+		.expect(200)
+})
